+++ kclib/_string_memops.c
@@ -8,7 +8,7 @@ void* memmove(void* s1, const void* s2, size_t n){
 	if (b1 == b2)
 		return b1;
 
-	for (uint32_t i = 0; i < n; i++){
+	for (size_t i = 0; i < n; i++){
 		b1[i] = b2[i];
 	}
 	return b2;
@@ -24,9 +24,9 @@ void* memcpy(void* restrict s1, const void* restrict s2, size_t n){
 	return (void*)s2;
 }
 
-void memset(void* ptr, uint32_t c, size_t n){
+void memset(void* ptr, int32_t c, size_t n){
 	register unsigned char* cptr = (unsigned char*)ptr;
 	register uint8_t chv = c;
-	for (register uint32_t i = 0; i < n; i++)
+	for (register size_t i = 0; i < n; i++)
 		cptr[i] = chv;
 }
